Database.Beam.Postgres.Migrate

Description

Migrations support for beam-postgres. See Database.Beam.Migrate for more information on beam migrations.

Documentation

data PgCommandSyntax

Representation of an arbitrary Postgres command. This is the combination of the command syntax (repesented by PgSyntax), as well as the type of command (represented by PgCommandType). The command type is necessary for us to know how to retrieve results from the database.

migrationBackend :: BeamMigrationBackend Postgres Pg

Top-level migration backend for use by beam-migrate tools

postgresDataTypeDeserializers :: BeamDeserializers Postgres

BeamDeserializers for postgres-specific types:

bytea

smallserial

serial

bigserial

tsvector

tsquery

text

json

jsonb

uuid

money

pgPredConverter :: HaskellPredicateConverter

Converts postgres DatabasePredicates to DatabasePredicates in the Haskell syntax. Allows automatic generation of Haskell schemas from postgres constraints.

getDbConstraints :: Connection -> IO [SomeDatabasePredicate]

getDbConstraintsForSchemas :: Maybe [String] -> Connection -> IO [SomeDatabasePredicate]

pgTypeToHs :: PgDataTypeSyntax -> Maybe HsDataType

Turn a PgDataTypeSyntax into the corresponding HsDataType. This is a best effort guess, and may fail on more exotic types. Feel free to send PRs to make this function more robust!

migrateScript :: MigrationSteps Postgres () a' -> [ByteString]

Turn a series of MigrationSteps into a line-by-line array of ByteStrings suitable for writing to a script.

writeMigrationScript :: FilePath -> MigrationSteps Postgres () a -> IO ()

Write the migration given by the MigrationSteps to a file.

pgDataTypeFromAtt :: ByteString -> Oid -> Maybe Int32 -> Maybe PgDataTypeSyntax

Postgres data types

tsquery :: DataType Postgres TsQuery

DataType for tsquery. See TsQuery for more information

tsvector :: DataType Postgres TsVector

DataType for tsvector. See TsVector for more information

text :: DataType Postgres Text

DataType for Postgres TEXT. characterLargeObject is also mapped to this data type

bytea :: DataType Postgres ByteString

DataType for Postgres BYTEA. binaryLargeObject is also mapped to this data type

unboundedArray :: forall a. Typeable a => DataType Postgres a -> DataType Postgres (Vector a)

DataType for a Postgres array without any bounds.

Note that array support in beam-migrate is still incomplete.

uuid :: DataType Postgres UUID

DataType for UUID columns. The pgCryptoGenRandomUUID function in the PgCrypto extension can be used to generate UUIDs at random.

money :: DataType Postgres PgMoney

DataType for MONEY columns.

json :: (ToJSON a, FromJSON a) => DataType Postgres (PgJSON a)

DataType for JSON. See PgJSON for more information

jsonb :: (ToJSON a, FromJSON a) => DataType Postgres (PgJSONB a)

DataType for JSONB. See PgJSON for more information

smallserial :: Integral a => DataType Postgres (SqlSerial a)

Postgres SERIAL data types. Automatically generates an appropriate DEFAULT clause and sequence

serial :: Integral a => DataType Postgres (SqlSerial a)

Postgres SERIAL data types. Automatically generates an appropriate DEFAULT clause and sequence

bigserial :: Integral a => DataType Postgres (SqlSerial a)

Postgres SERIAL data types. Automatically generates an appropriate DEFAULT clause and sequence

point :: DataType Postgres PgPoint

line :: DataType Postgres PgLine

lineSegment :: DataType Postgres PgLineSegment

box :: DataType Postgres PgBox

Orphan instances

BeamSqlBackendHasSerial Postgres

Instance details

Methods

genericSerial :: FieldReturnType 'True 'False Postgres (SqlSerial Int) a => Text -> a