Database.Beam.Postgres.PgCrypto

Description

The pgcrypto extension provides several cryptographic functions to Postgres. This module provides a beam-postgres extension to access this functionality. For an example of usage, see the documentation for PgExtensionEntity.

Documentation

data PgCrypto

Data type representing definitions contained in the pgcrypto extension

Each field maps closely to the underlying pgcrypto function, which are described in further detail in the pgcrypto manual.

Constructors

PgCrypto

Fields

pgCryptoDigestText :: forall ctxt s. LiftPg ctxt s (Text -> Text -> ByteString)

pgCryptoDigestBytes :: forall ctxt s. LiftPg ctxt s (ByteString -> Text -> ByteString)

pgCryptoHmacText :: forall ctxt s. LiftPg ctxt s (Text -> Text -> Text -> ByteString)

pgCryptoHmacBytes :: forall ctxt s. LiftPg ctxt s (ByteString -> Text -> Text -> ByteString)

pgCryptoCrypt :: forall ctxt s. LiftPg ctxt s (Text -> Text -> Text)

pgCryptoGenSalt :: forall ctxt s. LiftPg ctxt s (Text -> Maybe Int32 -> Text)

pgCryptoPgpSymEncrypt :: forall ctxt s. LiftPg ctxt s (Text -> Text -> Maybe Text -> ByteString)

pgCryptoPgpSymEncryptBytea :: forall ctxt s. LiftPg ctxt s (ByteString -> Text -> Maybe Text -> ByteString)

pgCryptoPgpSymDecrypt :: forall ctxt s. LiftPg ctxt s (ByteString -> Text -> Maybe Text -> Text)

pgCryptoPgpSymDecryptBytea :: forall ctxt s. LiftPg ctxt s (ByteString -> Text -> Maybe Text -> ByteString)

pgCryptoPgpPubEncrypt :: forall ctxt s. LiftPg ctxt s (Text -> ByteString -> Maybe Text -> ByteString)

pgCryptoPgpPubEncryptBytea :: forall ctxt s. LiftPg ctxt s (ByteString -> ByteString -> Maybe Text -> ByteString)

pgCryptoPgpPubDecrypt :: forall ctxt s. LiftPg ctxt s (ByteString -> ByteString -> Maybe Text -> Maybe Text -> Text)

pgCryptoPgpPubDecryptBytea :: forall ctxt s. LiftPg ctxt s (ByteString -> ByteString -> Maybe Text -> Maybe Text -> ByteString)

pgCryptoPgpKeyId :: forall ctxt s. LiftPg ctxt s (ByteString -> Text)

pgCryptoArmor :: forall ctxt s. PgExpr ctxt s ByteString -> Maybe (PgExpr ctxt s (Vector Text), PgExpr ctxt s (Vector Text)) -> PgExpr ctxt s Text

pgCryptoDearmor :: forall ctxt s. LiftPg ctxt s (Text -> ByteString)

pgCryptoGenRandomBytes :: forall ctxt s i. Integral i => PgExpr ctxt s i -> PgExpr ctxt s ByteString

pgCryptoGenRandomUUID :: forall ctxt s. PgExpr ctxt s UUID