Database.Beam.Postgres.Syntax

Description

Data types for Postgres syntax. Access is given mainly for extension modules. The types and definitions here are likely to change.

Documentation

data PgSyntaxF f where

Constructors

EmitByteString :: ByteString -> f -> PgSyntaxF f

EmitBuilder :: Builder -> f -> PgSyntaxF f

EscapeString :: ByteString -> f -> PgSyntaxF f

EscapeBytea :: ByteString -> f -> PgSyntaxF f

EscapeIdentifier :: ByteString -> f -> PgSyntaxF f

type PgSyntaxM = F PgSyntaxF

newtype PgSyntax

A piece of Postgres SQL syntax, which may contain embedded escaped byte and text sequences. PgSyntax composes monoidally, and may be created with emit, emitBuilder, escapeString, escapBytea, and escapeIdentifier.

Constructors

PgSyntax

Fields

buildPgSyntax :: PgSyntaxM ()

emit :: ByteString -> PgSyntax

emitBuilder :: Builder -> PgSyntax

escapeString :: ByteString -> PgSyntax

escapeBytea :: ByteString -> PgSyntax

escapeIdentifier :: ByteString -> PgSyntax

pgParens :: PgSyntax -> PgSyntax

nextSyntaxStep :: PgSyntaxF f -> f

data PgCommandSyntax

Representation of an arbitrary Postgres command. This is the combination of the command syntax (repesented by PgSyntax), as well as the type of command (represented by PgCommandType). The command type is necessary for us to know how to retrieve results from the database.

Constructors

PgCommandSyntax

Fields

pgCommandType :: PgCommandType

fromPgCommand :: PgSyntax

data PgCommandType

Constructors

PgCommandTypeQuery

PgCommandTypeDdl

PgCommandTypeDataUpdate

PgCommandTypeDataUpdateReturning

newtype PgSelectSyntax

IsSql92SelectSyntax for Postgres

Constructors

PgSelectSyntax

Fields

fromPgSelect :: PgSyntax

newtype PgSelectSetQuantifierSyntax

Constructors

PgSelectSetQuantifierSyntax

Fields

fromPgSelectSetQuantifier :: PgSyntax

newtype PgInsertSyntax

IsSql92InsertSyntax for Postgres

Constructors

PgInsertSyntax

Fields

fromPgInsert :: PgSyntax

newtype PgDeleteSyntax

IsSql92DeleteSyntax for Postgres

Constructors

PgDeleteSyntax

Fields

fromPgDelete :: PgSyntax

newtype PgUpdateSyntax

IsSql92UpdateSyntax for Postgres

Constructors

PgUpdateSyntax

Fields

fromPgUpdate :: PgSyntax

newtype PgExpressionSyntax

Constructors

PgExpressionSyntax

Fields

fromPgExpression :: PgSyntax

newtype PgFromSyntax

Constructors

PgFromSyntax

Fields

fromPgFrom :: PgSyntax

newtype PgTableNameSyntax

Constructors

PgTableNameSyntax

Fields

fromPgTableName :: PgSyntax

newtype PgComparisonQuantifierSyntax

Constructors

PgComparisonQuantifierSyntax

Fields

fromPgComparisonQuantifier :: PgSyntax

newtype PgExtractFieldSyntax

Constructors

PgExtractFieldSyntax

Fields

fromPgExtractField :: PgSyntax

newtype PgProjectionSyntax

Constructors

PgProjectionSyntax

Fields

fromPgProjection :: PgSyntax

newtype PgGroupingSyntax

Constructors

PgGroupingSyntax

Fields

fromPgGrouping :: PgSyntax

data PgOrderingSyntax

Constructors

PgOrderingSyntax

Fields

pgOrderingSyntax :: PgSyntax

pgOrderingNullOrdering :: Maybe PgNullOrdering

newtype PgValueSyntax

Constructors

PgValueSyntax

Fields

fromPgValue :: PgSyntax

newtype PgTableSourceSyntax

Constructors

PgTableSourceSyntax

Fields

fromPgTableSource :: PgSyntax

newtype PgFieldNameSyntax

Constructors

PgFieldNameSyntax

Fields

fromPgFieldName :: PgSyntax

newtype PgAggregationSetQuantifierSyntax

Constructors

PgAggregationSetQuantifierSyntax

Fields

fromPgAggregationSetQuantifier :: PgSyntax

newtype PgInsertValuesSyntax

Constructors

PgInsertValuesSyntax

Fields

fromPgInsertValues :: PgSyntax

newtype PgInsertOnConflictSyntax

Constructors

PgInsertOnConflictSyntax

Fields

fromPgInsertOnConflict :: PgSyntax

newtype PgInsertOnConflictTargetSyntax

Constructors

PgInsertOnConflictTargetSyntax

Fields

fromPgInsertOnConflictTarget :: PgSyntax

newtype PgConflictActionSyntax

Constructors

PgConflictActionSyntax

Fields

fromPgConflictAction :: PgSyntax

newtype PgCreateTableSyntax

Constructors

PgCreateTableSyntax

Fields

fromPgCreateTable :: PgSyntax

data PgTableOptionsSyntax

Constructors

PgTableOptionsSyntax PgSyntax PgSyntax

newtype PgColumnSchemaSyntax

Constructors

PgColumnSchemaSyntax

Fields

fromPgColumnSchema :: PgSyntax

data PgDataTypeSyntax

Constructors

PgDataTypeSyntax

Fields

pgDataTypeDescr :: PgDataTypeDescr

fromPgDataType :: PgSyntax

pgDataTypeSerialized :: BeamSerializedDataType

data PgColumnConstraintDefinitionSyntax

Constructors

PgColumnConstraintDefinitionSyntax

Fields

fromPgColumnConstraintDefinition :: PgSyntax

pgColumnConstraintDefinitionSerialized :: BeamSerializedConstraintDefinition

data PgColumnConstraintSyntax

Constructors

PgColumnConstraintSyntax

Fields

fromPgColumnConstraint :: PgSyntax

pgColumnConstraintSerialized :: BeamSerializedConstraint

newtype PgTableConstraintSyntax

Constructors

PgTableConstraintSyntax

Fields

fromPgTableConstraint :: PgSyntax

data PgMatchTypeSyntax

Constructors

PgMatchTypeSyntax

Fields

fromPgMatchType :: PgSyntax

pgMatchTypeSerialized :: BeamSerializedMatchType

data PgReferentialActionSyntax

Constructors

PgReferentialActionSyntax

Fields

fromPgReferentialAction :: PgSyntax

pgReferentialActionSerialized :: BeamSerializedReferentialAction

newtype PgAlterTableSyntax

Constructors

PgAlterTableSyntax

Fields

fromPgAlterTable :: PgSyntax

newtype PgAlterTableActionSyntax

Constructors

PgAlterTableActionSyntax

Fields

fromPgAlterTableAction :: PgSyntax

newtype PgAlterColumnActionSyntax

Constructors

PgAlterColumnActionSyntax

Fields

fromPgAlterColumnAction :: PgSyntax

newtype PgWindowFrameSyntax

Constructors

PgWindowFrameSyntax

Fields

fromPgWindowFrame :: PgSyntax

newtype PgWindowFrameBoundsSyntax

Constructors

PgWindowFrameBoundsSyntax

Fields

fromPgWindowFrameBounds :: PgSyntax

newtype PgWindowFrameBoundSyntax

Constructors

PgWindowFrameBoundSyntax

Fields

fromPgWindowFrameBound :: ByteString -> PgSyntax

data PgSelectLockingClauseSyntax

Constructors

PgSelectLockingClauseSyntax

Fields

pgSelectLockingClauseStrength :: PgSelectLockingStrength

pgSelectLockingTables :: [Text]

pgSelectLockingClauseOptions :: Maybe PgSelectLockingOptions

data PgSelectLockingStrength

Specifies the level of lock that will be taken against a row. See the manual section for more information.

Constructors

PgSelectLockingStrengthUpdate
UPDATE

PgSelectLockingStrengthNoKeyUpdate
NO KEY UPDATE

PgSelectLockingStrengthShare
SHARE

PgSelectLockingStrengthKeyShare
KEY SHARE

data PgSelectLockingOptions

Specifies how we should handle lock conflicts.

See the manual section for more information

Constructors

PgSelectLockingOptionsNoWait
NOWAIT. Report an error rather than waiting for the lock

PgSelectLockingOptionsSkipLocked
SKIP LOCKED. Rather than wait for a lock, skip the row instead

fromPgSelectLockingClause :: PgSelectLockingClauseSyntax -> PgSyntax

pgSelectStmt

Arguments

:: PgSelectTableSyntax

-> [PgOrderingSyntax]

-> Maybe Integer
LIMIT

-> Maybe Integer
OFFSET

-> Maybe PgSelectLockingClauseSyntax

-> PgSelectSyntax

defaultPgValueSyntax :: ToField a => a -> PgValueSyntax

data PgDataTypeDescr

Constructors

PgDataTypeDescrOid Oid (Maybe Int32)

PgDataTypeDescrDomain Text

data PgHasEnum

Constructors

PgHasEnum Text [Text]

pgCreateExtensionSyntax :: Text -> PgCommandSyntax

pgDropExtensionSyntax :: Text -> PgCommandSyntax

pgCreateEnumSyntax :: Text -> [PgValueSyntax] -> PgCommandSyntax

pgDropTypeSyntax :: Text -> PgCommandSyntax

pgSimpleMatchSyntax :: PgMatchTypeSyntax

pgSelectSetQuantifierDistinctOn :: [PgExpressionSyntax] -> PgSelectSetQuantifierSyntax

pgDataTypeJSON :: Value -> BeamSerializedDataType

pgTsQueryType :: PgDataTypeSyntax

pgTsVectorType :: PgDataTypeSyntax

pgJsonType :: PgDataTypeSyntax

pgJsonbType :: PgDataTypeSyntax

pgUuidType :: PgDataTypeSyntax

pgMoneyType :: PgDataTypeSyntax

pgTsQueryTypeInfo :: TypeInfo

pgTsVectorTypeInfo :: TypeInfo

Postgres TypeInfo for tsvector TODO Is the Oid stable from postgres instance to postgres instance?

pgByteaType :: PgDataTypeSyntax

pgTextType :: PgDataTypeSyntax

pgUnboundedArrayType :: PgDataTypeSyntax -> PgDataTypeSyntax

pgSerialType :: PgDataTypeSyntax

pgSmallSerialType :: PgDataTypeSyntax

pgBigSerialType :: PgDataTypeSyntax

pgPointType :: PgDataTypeSyntax

pgLineType :: PgDataTypeSyntax

pgLineSegmentType :: PgDataTypeSyntax

pgBoxType :: PgDataTypeSyntax

pgQuotedIdentifier :: Text -> PgSyntax

pgSepBy :: PgSyntax -> [PgSyntax] -> PgSyntax

pgDebugRenderSyntax :: PgSyntax -> IO ()

pgRenderSyntaxScript :: PgSyntax -> ByteString

pgBuildAction :: [Action] -> PgSyntax

pgBinOp :: ByteString -> PgExpressionSyntax -> PgExpressionSyntax -> PgExpressionSyntax

pgCompOp :: ByteString -> Maybe PgComparisonQuantifierSyntax -> PgExpressionSyntax -> PgExpressionSyntax -> PgExpressionSyntax

pgUnOp :: ByteString -> PgExpressionSyntax -> PgExpressionSyntax

pgPostFix :: ByteString -> PgExpressionSyntax -> PgExpressionSyntax

pgTestSyntax :: PgSyntax -> [PgSyntaxPrim]

data PostgresInaccessible